feat: add tamil department and OTP  mail

+++ controller/LibraryController/librarianController.js
@@ -1,12 +1,16 @@
-const librarian = require("./../../model/librarianModel")
+const { librarian } = require("./../../model/teacherModel")
 const asyncErrorHandler = require("./../../utils/asyncErrorHandler")
-
+const customError = require("./../../utils/customError")
 
 exports.createLibrarian = asyncErrorHandler( async (req,res,next) =>{
 
-    await librarian.create(req.body)
+    if( req?.userData?.role != "admin") throw new customError("Only admin can created a librarian")
+    req.body.route = "library"
+    req.body.role = "librarian"
 
-    req.status(201).json({
+    await librarian.create( req.body )
+
+    res.status(201).json({
         status:"Success",
         message:"Librarian successfully created"
     })
@@ -17,7 +21,7 @@ exports.getLibrarian = asyncErrorHandler(async (req,res,next) =>{
 
     let data = await librarian.find()
 
-    req.status(201).json({
+    res.status(200).json({
         status:"Success",
         length:data.length,
         data:{
@@ -29,13 +33,30 @@ exports.getLibrarian = asyncErrorHandler(async (req,res,next) =>{
 
 exports.deleteLibrarian = asyncErrorHandler(async (req,res,next) =>{
 
-    let id =  req.librarianId
+    if( req?.userData?.role != "admin") throw new customError("Only admin can created a librarian")
+
+    let id =  req.params.id
 
     await librarian.deleteOne({_id:id})
 
-    req.status(201).json({
+    res.status(204).json({
         status:"Success",
         message:"Successdfully librarian deleted"
+    })
+
+})
+
+exports.updateLibrarian = asyncErrorHandler(async (req,res,next) =>{
+
+    if( req?.userData?.role != "admin") throw new customError("Only admin can created a librarian")
+
+    let {id} =  req.body
+
+    await librarian.updateOne({_id:id},req.body)
+
+    res.status(201).json({
+        status:"Success",
+        message:"Successdfully librarian was updated"
     })
 
 })
